Aligned hero with grow, shrink, and basis
Applied grow, shrink, and basis to hero section divs. Set .header with
flex-grow/flex-shrink: 1; set .row with max-width and max-height 100%
and flex-shrink: 1; set .col-text and .col-img to flex, grow, shrink,
and flex-basis: 50%; set min-width and max-width of img; set flex-basis
and margin-bottom of .main-text and .secondary-text; added color and
font-size to button.